Allow specifying `.` for the target path to save to source folder
Currently the resulting image files are committed to a dedicated
folder specified by the `path` option.

This change adds the option to specify `.` as the value for `path`
in which case the rendered SVG images are stored in the same folder
as the source file it was generated from. This works for both
inline plantuml blocks as well as separate files.